requirements.txt
streamlit
pymupdf
camelot-py[cv]
opencv-python
sentence-transformers
faiss-cpu
torch
torchvision
transformers
pillow
pandas
numpy
tqdm
openai
